requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
backoff==2.2.1
bcrypt==5.0.0
bitsandbytes==0.42.0
brotli==1.2.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.2
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
ctranslate2==4.7.1
cycler==0.12.1
ddgs==9.10.0
decorator==5.2.1
distro==1.9.0
duckduckgo_search==8.1.1
durationpy==0.10
fake-useragent==2.2.0
fastapi==0.129.2
faster-whisper==1.2.1
ffmpy==1.0.0
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
googleapis-common-protos==1.72.0
gradio==6.6.0
gradio_client==2.1.0
groovy==0.1.2
groq==1.0.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==35.0.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.21.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
posthog==5.4.0
primp==1.1.2
protobuf==6.33.5
psutil==7.2.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
sentence-transformers==5.2.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soundfile==0.13.1
soxr==1.0.0
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.50.1
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
zipp==3.23.0
